+++ tests/Architecture/PHPat/Selector/Support/PHPUnitTestClassAnalysis.php
@@ -1,0 +1,97 @@
+<?php
+/**
+ * This code is licensed under the BSD 3-Clause License.
+ *
+ * Copyright (c) 2017, Maks Rafalko
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * * Redistributions of source code must retain the above copyright notice, this
+ *   list of conditions and the following disclaimer.
+ *
+ * * Redistributions in binary form must reproduce the above copyright notice,
+ *   this list of conditions and the following disclaimer in the documentation
+ *   and/or other materials provided with the distribution.
+ *
+ * * Neither the name of the copyright holder nor the names of its
+ *   contributors may be used to endorse or promote products derived from
+ *   this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+ * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+ * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+ * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+ * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+declare(strict_types=1);
+
+namespace Infection\Tests\Architecture\PHPat\Selector\Support;
+
+use PHPStan\BetterReflection\Reflection\Adapter\ReflectionAttribute;
+use PHPStan\Reflection\ClassReflection;
+use PHPUnit\Framework\Attributes\CoversNothing;
+use PHPUnit\Framework\Attributes\Group;
+use PHPUnit\Framework\TestCase;
+
+final class PHPUnitTestClassAnalysis
+{
+    private const string GROUP_NAME = 'integration';
+
+    public static function isPHPUnitTestCase(ClassReflection $classReflection): bool
+    {
+        return $classReflection->is(TestCase::class);
+    }
+
+    public static function hasCoversNothing(ClassReflection $classReflection): bool
+    {
+        return $classReflection->getNativeReflection()->getAttributes(CoversNothing::class) !== [];
+    }
+
+    public static function belongsToIntegrationGroup(ClassReflection $classReflection): bool
+    {
+        foreach (self::getAttributes($classReflection) as $groupAttribute) {
+            if (self::isIntegrationGroup($groupAttribute)) {
+                return true;
+            }
+        }
+
+        return false;
+    }
+
+    /**
+     * @see Group
+     */
+    private static function isIntegrationGroup(ReflectionAttribute $attribute): bool
+    {
+        if ($attribute->getName() !== Group::class) {
+            return false;
+        }
+
+        $arguments = $attribute->getArguments();
+        $groupName = $arguments[0] ?? $arguments['name'] ?? null;
+
+        return $groupName === self::GROUP_NAME;
+    }
+
+    /**
+     * @return iterable<ReflectionAttribute>
+     */
+    private static function getAttributes(ClassReflection $classReflection): iterable
+    {
+        $attributes = $classReflection->getNativeReflection()->getAttributes();
+
+        foreach ($attributes as $attribute) {
+            if ($attribute instanceof ReflectionAttribute) {
+                yield $attribute;
+            }
+        }
+    }
+}
