+++ tests/phpunit/Architecture/PHPat/Selector/ConcretePHPUnitTestClassTest.php
@@ -35,12 +35,14 @@ declare(strict_types=1);
 
 namespace Infection\Tests\Architecture\PHPat\Selector;
 
+use Infection\Tests\Architecture\PHPat\Selector\Support\PHPUnitTestClassAnalysis;
 use Infection\Tests\PhpParser\Visitor\BaseVisitorTestCase;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 
 #[CoversClass(ConcretePHPUnitTestClass::class)]
+#[CoversClass(PHPUnitTestClassAnalysis::class)]
 final class ConcretePHPUnitTestClassTest extends SelectorTestCase
 {
     /**
